+++ packages/backend/package.json
@@ -8,7 +8,7 @@
     "moment": "^2.24.0",
     "mysql": "^2.17.1",
     "node-fetch": "^2.4.1",
-    "nodemailer": "^6.3.0",
+    "nodemailer": "^6.3.1",
     "passport": "^0.4.0",
     "passport-local": "^1.0.0",
     "path": "^0.12.7",
