+++ packages/backend/password.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html lang="kr">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>냥터레스트 비밀번호 수정</title>
+</head>
+<body>
+        <div>
+            <label for="password">
+                비밀번호(6자이상 15자 이하)
+                <input type="password" name="password" id="password" placeholder="비밀번호입력"/>
+            </label>
+        </div>
+        <div>
+            <label for="password">
+                비밀번호확인(6자이상 15자 이하)
+                <input type="password" name="passwordCheck" id="passwordCheck" placeholder="입력한 비밀번호를 다시한번 입력해주세요"/>
+            </label>
+        </div>
+        <div>
+            <button type="button" class="submit-password" onclick="passwordModify()"></button>
+        </div>
+    <script>
+        const getUrlParams = ()=>{
+            var params = {};
+            window.location.search.replace(/[?&]+([^=&]+)=([^&]*)/gi, function(str, key, value) { params[key] = value; });
+            return params;
+        }
+
+        const passwordModify = (e) => {
+            const params = getUrlParams();
+            const passwordInfo = {
+                password: document.querySelector('#password').value,
+                passwordCheck: document.querySelector('#passwordCheck').value,
+                modifyToken: params.token
+            }
+
+            const jsonPasswordInfo = JSON.stringify(passwordInfo);
+
+            console.log(jsonPasswordInfo)
+            fetch('/account/password/modify',{
+                headers: {
+                'Accept': 'application/json',
+                'Content-Type': 'application/json'
+                },
+                method: "POST",
+                body: jsonPasswordInfo
+            })
+        }
+    </script>
+</body>
+</html>
